Refactoring and adding success callback to the sync process
Using a Queue to allow for linear processing of user data that can be chained
together. Now that there is a known completion point a callback can be given to
execute when the process is complete.

The dependency on `http` has been removed as well and instead the `UserSync`
module now operates on a stream. This allows the JSON data to come from a
number of sources including stdin.